添加一部分显示优化

+++ src/main/java/cn/qiuye/gtmoremachine/api/misc/BasicTransferData.java
@@ -2,6 +2,7 @@ package cn.qiuye.gtmoremachine.api.misc;
 
 import com.gregtechceu.gtceu.api.machine.MetaMachine;
 
+import java.math.BigInteger;
 import java.util.UUID;
 
-public record BasicTransferData(UUID UUID, long Throughput, MetaMachine machine) implements ITransferData {}
+public record BasicTransferData(UUID UUID, BigInteger Throughput, MetaMachine machine) implements ITransferData {}
